Aliment seems solved. if I delete a whole line and replace it for a new one
in the left/before it gets highlighted in red as deleted text, which is correct.

and on the right/after it gets Highlighted in green which is also correct.

but the position of it in before and after, makes it so that the after is below it, they should be aligned side by side in cases like this please.

when fixing this do make sure you preserve all other things, for instance any text below cases like this should continue to display as usual in before and after.

Please ensure that the changes I requested are applied through the composer, and most importantly, do not mess up any of the systems and logic that I have already worked hard to set up.